use a cube for the thread profile
Rather than use 2 cones for the profile,
use one cube turned up on its corners.
This involves some tricky trigonometry
to get the exact values of everything,
but it is all angles and lengths on the
unit cube.

Upsides:
* fewer lines of code
  * don't need to reset $fn
  * don't need to make 2 cones
* assymetric angles
  * easier to put on
  * harder to fall off

Downsides:
* hairy-looking trig math
* 1 long line of hairy-looking trig math
* need to flip outer tube
  to get the angles the right way around

Although now that I think about it,
I've always needed to flip the outer tube
so the threads run the right direction!