clarify how thread pitch is calculated
Technically, thread pitch should be calculated
as a function of turns per axial length. Maybe
I should switch to that in the future. In the
meanwhile, the pitch angle works fine, but I felt
I should not that it is calculated as a pitch
when wrapped around the inside diameter.