Don't use maximum length, just use tube length
Originally, ml would be the longest thing the tube
could hold, and the 2 halves would be ml/2 long.
The cap would add enough extra length for the two
parts to overlap.

The only thing that was done with ml was to divide
id by 2 to get part height. Everything else was done
with ph. So just use that, and rename it as Tube Length.