calculate thread pitch from diameter
rather than making it a configurable parameter.
The old default of 6mm was okay for very narrow
tubes, but too fine for larger tubes. Better to
have a good ratio and calculate it against the
parameter that the user actually knows (how wide
they need the tube to be) then making them have
to fiddle with another parameter.

Users who think the pitch is too coarse or too
fine can adjust the denominator in the fraction.